Rearrange bot tests and name them more in-line with standards
