+++ lib/auth/auth_with_roles_test.go
@@ -24,8 +24,6 @@ import (
 	"testing"
 	"time"
 
-	"golang.org/x/crypto/ssh"
-
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/client/proto"
 	"github.com/gravitational/teleport/api/constants"
@@ -39,7 +37,6 @@ import (
 	"github.com/gravitational/teleport/lib/fixtures"
 	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/tlsca"
-	"github.com/gravitational/teleport/lib/utils"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
@@ -123,160 +120,6 @@ func TestLocalUserCanReissueCerts(t *testing.T) {
 			require.WithinDuration(t, start.Add(test.expiresIn), x509.NotAfter, 1*time.Second)
 		})
 	}
-}
-
-func renewBotCerts(
-	srv *TestTLSServer, tlsCert tls.Certificate, botUser string,
-	publicKey []byte, privateKey []byte,
-) (*Client, *proto.Certs, tls.Certificate, error) {
-	client := srv.NewClientWithCert(tlsCert)
-
-	certs, err := client.GenerateUserCerts(context.Background(), proto.UserCertsRequest{
-		PublicKey: publicKey,
-		Username:  botUser,
-		Expires:   time.Now().Add(1 * time.Hour),
-	})
-	if err != nil {
-		return nil, nil, tls.Certificate{}, trace.Wrap(err)
-	}
-
-	// Make sure to overwrite tlsCert with the new certs.
-	tlsCert, err = tls.X509KeyPair(certs.TLS, privateKey)
-	if err != nil {
-		return nil, nil, tls.Certificate{}, trace.Wrap(err)
-	}
-
-	return client, certs, tlsCert, nil
-}
-
-// TestBotCertificateGenerationCheck ensures bot cert generation checks work
-// in ordinary conditions, with several rapid renewals.
-func TestBotCertificateGenerationCheck(t *testing.T) {
-	t.Parallel()
-	srv := newTestTLSServer(t)
-
-	_, err := CreateRole(context.Background(), srv.Auth(), "example", types.RoleSpecV5{})
-	require.NoError(t, err)
-
-	// Create a new bot.
-	bot, err := srv.Auth().createBot(context.Background(), &proto.CreateBotRequest{
-		Name:  "test",
-		Roles: []string{"example"},
-	})
-	require.NoError(t, err)
-
-	privateKey, publicKey, err := native.GenerateKeyPair()
-	require.NoError(t, err)
-	sshPrivateKey, err := ssh.ParseRawPrivateKey(privateKey)
-	require.NoError(t, err)
-	tlsPublicKey, err := tlsca.MarshalPublicKeyFromPrivateKeyPEM(sshPrivateKey)
-	require.NoError(t, err)
-
-	certs, err := Register(RegisterParams{
-		Token: bot.TokenID,
-		ID: IdentityID{
-			Role: types.RoleBot,
-		},
-		Servers:      []utils.NetAddr{*utils.MustParseAddr(srv.Addr().String())},
-		PublicTLSKey: tlsPublicKey,
-		PublicSSHKey: publicKey,
-	})
-	require.NoError(t, err)
-
-	tlsCert, err := tls.X509KeyPair(certs.TLS, privateKey)
-	require.NoError(t, err)
-
-	// Renew the cert a bunch of times.
-	for i := 0; i < 10; i++ {
-		_, certs, tlsCert, err = renewBotCerts(srv, tlsCert, bot.UserName, publicKey, privateKey)
-		require.NoError(t, err)
-
-		// Parse the Identity
-		impersonatedTLSCert, err := tlsca.ParseCertificatePEM(certs.TLS)
-		require.NoError(t, err)
-		impersonatedIdent, err := tlsca.FromSubject(impersonatedTLSCert.Subject, impersonatedTLSCert.NotAfter)
-		require.NoError(t, err)
-
-		// Cert must be renewable.
-		require.True(t, impersonatedIdent.Renewable)
-		require.False(t, impersonatedIdent.DisallowReissue)
-
-		// Initial certs have generation=1 and we start the loop with a renewal, so add 2
-		require.Equal(t, uint64(i+2), impersonatedIdent.Generation)
-	}
-}
-
-// TestBotCertificateGenerationStolen simulates a stolen renewable certificate
-// where a generation check is expected to fail.
-func TestBotCertificateGenerationStolen(t *testing.T) {
-	t.Parallel()
-	srv := newTestTLSServer(t)
-
-	_, err := CreateRole(context.Background(), srv.Auth(), "example", types.RoleSpecV5{})
-	require.NoError(t, err)
-
-	// Create a new bot.
-	bot, err := srv.Auth().createBot(context.Background(), &proto.CreateBotRequest{
-		Name:  "test",
-		Roles: []string{"example"},
-	})
-	require.NoError(t, err)
-
-	privateKey, publicKey, err := native.GenerateKeyPair()
-	require.NoError(t, err)
-	sshPrivateKey, err := ssh.ParseRawPrivateKey(privateKey)
-	require.NoError(t, err)
-	tlsPublicKey, err := tlsca.MarshalPublicKeyFromPrivateKeyPEM(sshPrivateKey)
-	require.NoError(t, err)
-
-	certs, err := Register(RegisterParams{
-		Token: bot.TokenID,
-		ID: IdentityID{
-			Role: types.RoleBot,
-		},
-		Servers:      []utils.NetAddr{*utils.MustParseAddr(srv.Addr().String())},
-		PublicTLSKey: tlsPublicKey,
-		PublicSSHKey: publicKey,
-	})
-	require.NoError(t, err)
-
-	tlsCert, err := tls.X509KeyPair(certs.TLS, privateKey)
-	require.NoError(t, err)
-
-	// Renew the certs once (e.g. this is the actual bot process)
-	_, certsReal, _, err := renewBotCerts(srv, tlsCert, bot.UserName, publicKey, privateKey)
-	require.NoError(t, err)
-
-	// Check the generation, it should be 2.
-	impersonatedTLSCert, err := tlsca.ParseCertificatePEM(certsReal.TLS)
-	require.NoError(t, err)
-	impersonatedIdent, err := tlsca.FromSubject(impersonatedTLSCert.Subject, impersonatedTLSCert.NotAfter)
-	require.NoError(t, err)
-	require.Equal(t, uint64(2), impersonatedIdent.Generation)
-
-	// Meanwhile, the initial set of certs was stolen. Let's try to renew those.
-	_, _, _, err = renewBotCerts(srv, tlsCert, bot.UserName, publicKey, privateKey)
-	require.Error(t, err)
-	require.True(t, trace.IsAccessDenied(err))
-
-	// The user should now be locked.
-	locks, err := srv.Auth().GetLocks(context.Background(), true, types.LockTarget{
-		User: "bot-test",
-	})
-	require.NoError(t, err)
-	require.NotEmpty(t, locks)
-}
-
-// TestBotNoRoles attempts to create a bot with an empty role list.
-func TestBotNoRoles(t *testing.T) {
-	t.Parallel()
-	srv := newTestTLSServer(t)
-
-	// Create a new bot without roles specified. This should fail.
-	_, err := srv.Auth().createBot(context.Background(), &proto.CreateBotRequest{
-		Name: "test",
-	})
-	require.True(t, trace.IsBadParameter(err))
 }
 
 // TestSSOUserCanReissueCert makes sure that SSO user can reissue certificate
@@ -574,7 +417,8 @@ func TestOIDCAuthRequest(t *testing.T) {
 					Roles: []string{"access"},
 				},
 			},
-		}}
+		},
+	}
 
 	tests := []struct {
 		desc               string
@@ -898,7 +742,6 @@ func TestSSODiagnosticInfo(t *testing.T) {
 	info, err = clientPriv.GetSSODiagnosticInfo(ctx, types.KindSAML, "ABC123")
 	require.NoError(t, err)
 	require.Equal(t, &infoCreate, info)
-
 }
 
 func TestGenerateUserCertsWithRoleRequest(t *testing.T) {
